Move initials creation logic to utility
This will enable easy unit testing, re-use, etc. going forward.
